+++ src/java/src/main/java/com/dnanexus/DXJob.java
@@ -18,16 +18,15 @@ package com.dnanexus;
 
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Sets;
 
 /**
@@ -40,93 +39,131 @@ public final class DXJob extends DXExecution {
      * this object was created.
      */
     public final static class Describe extends DXExecution.Describe {
-        private final DescribeResponseHash describeOutput;
-        private final DXEnvironment env;
 
-        // TODO: startedRunning
-        // TODO: stoppedRunning
-        // TODO: originJob
-        // TODO: stateTransitions
-        // TODO: function
-        // TODO: tags
-        // TODO: properties
+        private final DescribeResponseHash describeOutput;
+
         // TODO: systemRequirements
         // TODO: executionPolicy
         // TODO: instanceType
-        // TODO: dependsOn
-        // TODO: failure*
-        // TODO: isFree
-        // TODO: applet/app
-        // TODO: resources
-        // TODO: projectCache
-
-        // TODO: plus the common fields in DXExecution.Describe
+        // TODO: failureFrom, failureReports
+        // TODO: provide accessors for dependsOn
+        // TODO: provide accessor for app (there is no DXApp class right now)
 
         @VisibleForTesting
         Describe(DescribeResponseHash describeOutput, DXEnvironment env) {
+            super(describeOutput, env);
             this.describeOutput = describeOutput;
-            this.env = env;
         }
 
         /**
-         * Returns the date at which the job was created.
+         * Returns the applet of the job if it is running an applet, or null otherwise.
          *
-         * @return the job's creation date
+         * @return applet or null
          */
-        public Date getCreationDate() {
-            return new Date(describeOutput.created);
-        }
-
-        /**
-         * Returns the details (user-suppled metadata) of the job.
-         *
-         * @param valueType class to deserialize to
-         *
-         * @return the job's details
-         */
-        public <T> T getDetails(Class<T> valueType) {
-            return DXJSON.safeTreeToValue(describeOutput.details, valueType);
-        }
-
-        @Override
-        public String getId() {
-            return describeOutput.id;
-        }
-
-        /**
-         * Returns the date at which the job was last modified.
-         *
-         * @return the job's modification date
-         */
-        public Date getModifiedDate() {
-            return new Date(describeOutput.modified);
-        }
-
-        @Override
-        public String getName() {
-            return describeOutput.name;
-        }
-
-        @Override
-        public <T> T getOutput(Class<T> outputClass) {
-            return DXJSON.safeTreeToValue(describeOutput.output, outputClass);
-        }
-
-        /**
-         * Returns the job's parent job, or {@code null} if the job is an origin job.
-         *
-         * @return {@code DXJob} for parent job
-         */
-        public DXJob getParentJob() {
-            if (describeOutput.parentJob == null) {
+        public DXApplet getApplet() {
+            if (describeOutput.applet == null) {
                 return null;
             }
-            return new DXJob(describeOutput.parentJob, env);
+            return DXApplet.getInstanceWithEnvironment(describeOutput.applet, env);
         }
 
+        /**
+         * Returns a detailed message describing why the job failed, or null if the job is not in a
+         * failing or failed state.
+         *
+         * @return detailed failure message, or null
+         */
+        public String getFailureMessage() {
+            return describeOutput.failureMessage;
+        }
+
+        /**
+         * Returns a short String describing why the job failed, or null if the job is not in a
+         * failing or failed state.
+         *
+         * @return short failure reason, or null
+         */
+        public String getFailureReason() {
+            return describeOutput.failureReason;
+        }
+
+        /**
+         * Returns the name of the function (entry point) that the job is running.
+         *
+         * @return function name
+         */
+        public String getFunction() {
+            return describeOutput.function;
+        }
+
+        /**
+         * Returns the closest ancestor job whose parentJob is null. This is the nearest job that
+         * was run by a user directly or was run as a stage in an analysis.
+         *
+         * @return origin job
+         */
+        public DXJob getOriginJob() {
+            return DXJob.getInstanceWithEnvironment(describeOutput.originJob, env);
+        }
+
+        /**
+         * Returns the output of the job, deserialized to the specified class, or null if no output
+         * hash is available. Note that this field is not guaranteed to be complete until the job
+         * has finished.
+         *
+         * <p>
+         * A partial output hash is available as soon as the job reaches state "waiting_on_output".
+         * At this time the output may contain unresolved job-based object references. These
+         * references will be resolved by the time the job reaches state "done".
+         * </p>
+         *
+         * @param outputClass class to deserialize to
+         *
+         * @return output object or null
+         */
         @Override
-        public Map<String, String> getProperties() {
-            return ImmutableMap.copyOf(describeOutput.properties);
+        public <T> T getOutput(Class<T> outputClass) {
+            // We don't do anything different here except for providing the job-specific
+            // caveats in the docstring above.
+            return super.getOutput(outputClass);
+        }
+
+        /**
+         * Returns the project cache container for the job if it is running an app, or null
+         * otherwise.
+         *
+         * @return project cache container or null
+         */
+        public DXContainer getProjectCache() {
+            if (describeOutput.projectCache == null) {
+                return null;
+            }
+            return DXContainer.getInstanceWithEnvironment(describeOutput.projectCache, env);
+        }
+
+        /**
+         * Returns the resources container for the job if it is running an app, or null otherwise.
+         *
+         * @return resources container or null
+         */
+        public DXContainer getResources() {
+            if (describeOutput.resources == null) {
+                return null;
+            }
+            return DXContainer.getInstanceWithEnvironment(describeOutput.resources, env);
+        }
+
+        /**
+         * Returns the date that the job started running, or null if the job has not started running
+         * yet.
+         *
+         * @return start date or null
+         */
+        public Date getStartDate() {
+            if (describeOutput.startedRunning == null) {
+                return null;
+            }
+            return new Date(describeOutput.startedRunning);
         }
 
         /**
@@ -138,9 +175,50 @@ public final class DXJob extends DXExecution {
             return describeOutput.state;
         }
 
-        @Override
-        public List<String> getTags() {
-            return ImmutableList.copyOf(describeOutput.tags);
+        /**
+         * Returns a list of the state transitions for the job, showing each state that the job
+         * reached and at what time. Every job implicitly starts in state "idle" (which does not
+         * appear in the state transition list).
+         *
+         * @return List of state transitions
+         */
+        public List<StateTransition> getStateTransitions() {
+            return ImmutableList.copyOf(describeOutput.stateTransitions);
+        }
+
+        /**
+         * Returns the date that the job stopped running, or null if the job has not stopped running
+         * yet.
+         *
+         * @return stop date or null
+         */
+        public Date getStopDate() {
+            if (describeOutput.stoppedRunning == null) {
+                return null;
+            }
+            return new Date(describeOutput.stoppedRunning);
+        }
+
+        /**
+         * Returns whether the job will be charged to the billed entity (billTo).
+         *
+         * <p>
+         * Usually the job will be free if the job has failed for reasons (see
+         * {@link #getFailureReason()}) that are generally indicative of some system error rather
+         * than of user error.
+         * </p>
+         *
+         * @return true if the job is free
+         *
+         * @throws IllegalStateException if the requesting user does not have permission to view the
+         *         pricing model of the billed entity (billTo), or the price of the execution has
+         *         not been finalized
+         */
+        public boolean isFree() {
+            if (describeOutput.isFree == null) {
+                throw new IllegalStateException("isFree is not available");
+            }
+            return describeOutput.isFree;
         }
     }
 
@@ -150,28 +228,37 @@ public final class DXJob extends DXExecution {
      */
     @VisibleForTesting
     @JsonIgnoreProperties(ignoreUnknown = true)
-    final static class DescribeResponseHash {
-        @JsonProperty
-        private String id;
-        @JsonProperty
-        private String name;
-        @JsonProperty
-        private Long created;
-        @JsonProperty
-        private Long modified;
-        @JsonProperty
-        private String parentJob;
+    final static class DescribeResponseHash extends DXExecution.DescribeResponseHash {
         @JsonProperty
         private JobState state;
         @JsonProperty
-        private JsonNode details;
+        private Long startedRunning;
         @JsonProperty
-        private List<String> tags;
+        private Long stoppedRunning;
         @JsonProperty
-        private Map<String, String> properties;
-
+        private String originJob;
         @JsonProperty
-        private JsonNode output;
+        private List<StateTransition> stateTransitions;
+        @JsonProperty
+        private String function;
+        @SuppressWarnings("unused")
+        @JsonProperty
+        private List<String> dependsOn;
+        @JsonProperty
+        private String failureReason;
+        @JsonProperty
+        private String failureMessage;
+        @JsonProperty
+        private Boolean isFree;
+        @JsonProperty
+        private String applet;
+        @SuppressWarnings("unused")
+        @JsonProperty
+        private String app;
+        @JsonProperty
+        private String resources;
+        @JsonProperty
+        private String projectCache;
     }
 
     /**
@@ -179,6 +266,82 @@ public final class DXJob extends DXExecution {
      */
     @JsonIgnoreProperties(ignoreUnknown = true)
     private static class JobTerminateResponse {}
+
+    /**
+     * A event where a job transitioned from one state to another.
+     */
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    public final static class StateTransition {
+        @JsonProperty
+        private JobState newState;
+        @JsonProperty
+        private Long setAt;
+
+        @SuppressWarnings("unused")
+        private StateTransition() {
+            // No-arg constructor for JSON deserialization
+        }
+
+        @VisibleForTesting
+        StateTransition(JobState newState, long setAt) {
+            this.newState = newState;
+            this.setAt = setAt;
+        }
+
+        /* (non-Javadoc)
+         * @see java.lang.Object#equals(java.lang.Object)
+         */
+        @Override
+        public boolean equals(Object obj) {
+            if (this == obj) {
+                return true;
+            }
+            if (obj == null) {
+                return false;
+            }
+            if (!(obj instanceof StateTransition)) {
+                return false;
+            }
+            StateTransition other = (StateTransition) obj;
+            if (newState != other.newState) {
+                return false;
+            }
+            if (setAt == null) {
+                if (other.setAt != null) {
+                    return false;
+                }
+            } else if (!setAt.equals(other.setAt)) {
+                return false;
+            }
+            return true;
+        }
+
+        /**
+         * Returns the state of the job after the transition.
+         *
+         * @return new job state
+         */
+        public JobState getNewState() {
+            return newState;
+        }
+
+        /**
+         * Returns the date at which the new state became effective.
+         *
+         * @return transition time
+         */
+        public Date getSetAt() {
+            return new Date(setAt);
+        }
+
+        /* (non-Javadoc)
+         * @see java.lang.Object#hashCode()
+         */
+        @Override
+        public int hashCode() {
+            return Objects.hashCode(newState, setAt);
+        }
+    }
 
     private static final Set<JobState> unsuccessfulJobStates = Sets.immutableEnumSet(
             JobState.FAILED, JobState.TERMINATED);
@@ -251,7 +414,8 @@ public final class DXJob extends DXExecution {
     @Override
     public Describe getCachedDescribe() {
         this.checkCachedDescribeAvailable();
-        return new Describe(DXJSON.safeTreeToValue(this.cachedDescribe, DescribeResponseHash.class), this.env);
+        return new Describe(
+                DXJSON.safeTreeToValue(this.cachedDescribe, DescribeResponseHash.class), this.env);
     }
 
     @Override